move CosseratRodIO to utils
