+++ sopht_simulator/immersed_body/cosserat_rod/__init__.py
@@ -7,4 +7,3 @@ from sopht_simulator.immersed_body.cosserat_rod.cosserat_rod_forcing_grids impor
 from sopht_simulator.immersed_body.cosserat_rod.cosserat_rod_flow_interaction import (
     CosseratRodFlowInteraction,
 )
-from sopht_simulator.immersed_body.cosserat_rod.CosseratRodIO import CosseratRodIO
