+++ sopht_simulator/utils/__init__.py
@@ -2,5 +2,5 @@ from .lab_cmap import lab_cmap
 from .plot_field import create_figure_and_axes, save_and_clear_fig
 from .image_utils import make_video_from_image_series
 from .grid_utils import VectorField
-from .io_utils import make_dir_and_transfer_h5_data
+from .io_utils import make_dir_and_transfer_h5_data, CosseratRodIO
 from .rod_viz import plot_video_of_rod_surface
